+++ benchopt/benchmark.py
@@ -11,7 +11,7 @@ from joblib.externals import cloudpickle
 from .config import get_setting
 from .base import BaseSolver, BaseDataset
 
-from .utils.dynamic_modules import _load_class_from_module
+from .utils.dynamic_modules import _load_class_from_module, FailedImport
 from .utils.parametrized_name_mixin import product_param
 
 from .utils.terminal_output import colorify
@@ -444,7 +444,8 @@ class Benchmark:
             requirements["gpu"] instead of requirements["cpu"].
         """
         # Collect all classes matching one of the patterns
-        print("Collecting packages...", end='', flush=True)
+        print("Collecting packages:")
+        exit_code = 0
 
         check_installs, missings = [], []
         objective = self.get_benchmark_objective()
@@ -454,7 +455,7 @@ class Benchmark:
         if missing_deps:
             raise AttributeError(
                 "Could not find dependencies in objective.py while it is not "
-                f"importable. {MISSING_DEPS_MSG}"
+                f"importable.\n\n{MISSING_DEPS_MSG}"
             )
 
         if len(shell_install_scripts) > 0 or len(conda_reqs) > 0:
@@ -486,11 +487,14 @@ class Benchmark:
             f"- {klass.name}" for klass in check_installs
         ])
         if len(list_install) == 0:
-            self.check_missing(missings)
-            print("All required solvers are already installed.")
+            exit_code = self.check_missing(missings)
+            print("No new requirements installed")
             if download:
-                self.download_all_data(include_datasets, env_name, quiet)
-            return
+                exit_code = max(
+                    exit_code,
+                    self.download_all_data(include_datasets, env_name, quiet)
+                )
+            return exit_code
 
         print(f"Installing required packages for:\n{list_install}\n...",
               end='', flush=True)
@@ -517,7 +521,7 @@ class Benchmark:
             elif not cls_success:
                 not_installed.add(klass.name)
 
-        self.check_missing(missings)
+        exit_code = self.check_missing(missings)
 
         # If one failed, raise a warning to explain how to see the install
         # errors.
@@ -533,14 +537,18 @@ class Benchmark:
             print(colorify(f" done (missing deps: {not_installed})", YELLOW))
 
         if download:
-            self.download_all_data(include_datasets, env_name, quiet)
+            exit_code = max(
+                exit_code,
+                self.download_all_data(include_datasets, env_name, quiet)
+            )
+        return exit_code
 
     def download_all_data(self, datasets, env_name, quiet):
         if len(datasets) == 0:
-            return
+            return 0
         cmd = f"benchopt check-data {self.benchmark_dir} -d "
         cmd += " -d ".join(d.name for d in datasets)
-        _run_shell_in_conda_env(
+        return _run_shell_in_conda_env(
             cmd, env_name=env_name, raise_on_error=True, capture_stdout=False
         )
 
@@ -548,22 +556,38 @@ class Benchmark:
         # Check that classes not importable, with no requirements, only depends
         # on global requirements specified in Objective.requirements.
         # Otherwise, we raise a comprehensible error.
-        if len(missings) > 0:
-            # Format the list of classes missing requirements.
-            cls_types = {'Solver': [], 'Dataset': []}
-            for klass in missings:
-                cls_type = klass.__base__.__name__.replace("Base", "")
-                cls_types[cls_type].append(klass.name)
-            cls_types = {
-                k: f'{cls_type}\n' + '\n'.join([f'- {c}' for c in v])
-                for k, v in cls_types.items() if len(v) > 0
-            }
-            missing_cls = '\n'.join(cls_types.values())
+        if len(missings) == 0:
+            return 0
 
-            raise AttributeError(
-                f"Could not find dependencies for the following classes while "
-                f"they are not importable:\n{missing_cls}\n{MISSING_DEPS_MSG}"
-            )
+        # Format the list of classes missing requirements.
+        cls_types = {'Solver': [], 'Dataset': []}
+        for klass in missings:
+            cls_type = klass.__base__.__name__.replace("Base", "")
+            try:
+                # Check for invalid install_cmd
+                hasattr(klass, "install_cmd")
+                # Check for invalid requirements
+                if not hasattr(klass, "requirements"):
+                    reason = "no requirements"
+                else:
+                    reason = "incomplete requirements"
+            except ValueError as e:
+                if "install_cmd" in str(e):
+                    reason = "invalid install_cmd"
+                else:
+                    reason = "invalid requirements"
+            cls_types[cls_type].append(f"{klass.name} ({reason})")
+        cls_types = {
+            k: f'{cls_type}\n' + '\n'.join([f'- {c}' for c in v])
+            for k, v in cls_types.items() if len(v) > 0
+        }
+        missing_cls = '\n'.join(cls_types.values())
+
+        print(
+            f"Could not find dependencies for the following classes while "
+            f"they are not importable:\n{missing_cls}\n\n{MISSING_DEPS_MSG}"
+        )
+        return 1
 
     def get_all_runs(self, solvers=None, forced_solvers=None,
                      datasets=None, objectives=None, terminal=None):
@@ -817,17 +841,21 @@ def _check_patterns(all_classes, patterns, name_type='dataset',
         raise TypeError()
     patterns = [p for q in patterns for p in preprocess_patterns(q)]
 
-    # Check that the provided patterns match at least one dataset and pair the
-    # matching clas with the selector.
+    # Check that each provided pattern matches at least one dataset and pair
+    # the matching class with the selector.
     matched, invalid_patterns = [], []
     for p, args, kwargs in patterns:
-        matched += [
+        matched_cls = [
             (cls, (args, kwargs))
             for cls in all_classes
             if is_matched(cls.name, [p])
         ]
-        if len(matched) == 0:
+        if len(matched_cls) == 0:
             invalid_patterns.append(p)
+        matched.extend([
+            (cls, p) if not isinstance(cls, FailedImport) else (cls, ([], {}))
+            for cls, p in matched_cls
+        ])
 
     # If some patterns did not matched any class, raise an error
     if len(invalid_patterns) > 0:
@@ -844,7 +872,7 @@ def _check_patterns(all_classes, patterns, name_type='dataset',
     for cls, (args, kwargs) in matched:
         param_names = [p.strip() for k in cls.parameters for p in k.split(',')]
         if len(args) != 0:
-            if len(cls.parameters) > 1:
+            if len(param_names) > 1:
                 raise ValueError(
                     f"Ambiguous positional parameter for {cls.name}."
                 )
@@ -852,7 +880,13 @@ def _check_patterns(all_classes, patterns, name_type='dataset',
                 raise ValueError(
                     f"Both positional and keyword parameters for {cls.name}."
                 )
-            kwargs = {list(cls.parameters.keys())[0]: args}
+            elif len(param_names) == 0:
+                raise ValueError(
+                    f"Positional parameter provided for {cls.name} which has"
+                    " no parameter."
+                )
+            # Use the single parameter name for this class.
+            kwargs = {param_names[0]: args}
         else:
             bad_params = [
                 p.strip() for k in kwargs for p in k.split(',')