+++ benchopt/tests/test_stopping_criterion.py
@@ -151,7 +151,7 @@ def test_key_to_monitor_objective(no_debug_log, criterion_class, strategy):
                 .split()], standalone_mode=False)
 
     out.check_output('test-solver', 5)
-    out.check_output('done', 1)
+    out.check_output('test-solver: done', 1)
 
 
 @pytest.mark.parametrize('strategy', SAMPLING_STRATEGIES)
@@ -330,4 +330,4 @@ def test_objective_equals_zero(no_debug_log):
                 *('-s test-solver -d test-dataset --no-plot -n 0').split()],
                 standalone_mode=False)
 
-    out.check_output('done', 1)
+    out.check_output('test-solver: done', 1)