+++ benchopt/utils/dynamic_modules.py
@@ -1,5 +1,6 @@
 """Utilities to load classes and module from filenames and class names.
 """
+from abc import ABCMeta
 import ast
 import sys
 import hashlib
@@ -11,6 +12,7 @@ from joblib.externals import cloudpickle
 
 from .dependencies_mixin import DependenciesMixin
 from .safe_import import safe_import_context
+from .class_property import classproperty
 
 SKIP_IMPORT = False
 
@@ -26,6 +28,16 @@ def _unskip_import():
     """Helper to reenable imports in tests."""
     global SKIP_IMPORT
     SKIP_IMPORT = False
+
+
+class FailedImport(ABCMeta):
+    """MetaClass for fake classes mimicking components that cannot be imported.
+
+    This class should be used as a metaclass so that the class type can be
+    checked and to have an informative representation.
+    """
+    def __repr__(cls):
+        return f"<FailedImport: {cls.cls_name}({cls.name})|| Exc: [{cls.exc}]>"
 
 
 def _get_module_from_file(module_filename, benchmark_dir=None):
@@ -96,15 +108,18 @@ def _load_class_from_module(benchmark_dir, module_filename, class_name):
             Solver=BaseSolver, Dataset=BaseDataset, Objective=BaseObjective
         )[class_name]
 
-        class FailedImport(base_cls, ParametrizedNameMixin, DependenciesMixin):
+        class klass(base_cls, ParametrizedNameMixin, DependenciesMixin,
+                    metaclass=FailedImport):
             "Object for the class list that raises error if used."
 
-            _set_cls_attr_from_ast(module_filename, class_name, locals())
             exc = e
+            _error_displayed = False
+            _set_cls_attr_from_ast(module_filename, class_name, locals())
+            cls_name = class_name
 
             @classmethod
             def is_installed(cls, env_name=None, raise_on_not_installed=False,
-                             **kwargs):
+                             quiet=False, **kwargs):
                 if env_name is not None:
                     return super().is_installed(
                         env_name=env_name,
@@ -114,15 +129,15 @@ def _load_class_from_module(benchmark_dir, module_filename, class_name):
                 if not SKIP_IMPORT:
                     if raise_on_not_installed:
                         raise cls.exc
-                    print(
-                        f"Failed to import {class_name} from "
-                        f"{module_filename}. Please fix the following "
-                        "error to use this file with benchopt:\n"
-                        f"{tb_to_print}"
-                    )
+                    if not cls._error_displayed and not quiet:
+                        print(
+                            f"Failed to import {class_name} from "
+                            f"{module_filename}. Please fix the following "
+                            "error to use this file with benchopt:\n"
+                            f"{tb_to_print}"
+                        )
+                        cls._error_displayed = True
                 return False
-
-        klass = FailedImport
 
     # Store the info to easily reload the class and check it is installed
     klass._module_filename = module_filename.resolve()
@@ -197,11 +212,18 @@ def _reconstruct_class(
 
 def _set_cls_attr_from_ast(module_file, cls_name, ctx):
     module = ast.parse(module_file.read_text())
+    name = f"{module_file.stem}"
 
     cls_list = [node for node in module.body if isinstance(node, ast.ClassDef)
                 and node.name == cls_name]
     if not cls_list:
-        raise ValueError(f"Could not find {cls_name} in module {module_file}.")
+        exc = ValueError(
+            f"Could not find {cls_name} in module {module_file}."
+        )
+        exc.__cause__ = ctx['exc']
+        ctx['exc'] = exc
+        ctx['name'] = name
+        return
     cls = cls_list[0]
 
     known_methods = [
@@ -218,12 +240,37 @@ def _set_cls_attr_from_ast(module_file, cls_name, ctx):
     for node in cls.body:
         if isinstance(node, ast.Assign):
             for target in node.targets:
-                if target.id == "requirements":
-                    ctx['requirements'] = ast.literal_eval(node.value)
-                elif target.id == "name":
-                    ctx['name'] = ast.literal_eval(node.value)
-                elif target.id == "install_cmd":
-                    ctx['install_cmd'] = ast.literal_eval(node.value)
+                if target.id in ["name", "requirements", "install_cmd"]:
+                    try:
+                        ctx[target.id] = ast.literal_eval(node.value)
+                    except Exception:
+                        if target.id == "name":
+                            exc = ValueError(
+                                f"Could not evaluate the name of the class "
+                                f"{cls_name} in module {module_file}.\n"
+                                f"The name should be a string literal."
+                            )
+                            exc.__cause__ = ctx['exc']
+                            ctx['exc'] = exc
+                            ctx['name'] = name
+                        else:
+                            msg = (
+                                f"Could not evaluate statically '{target.id}' "
+                                f"of the class {cls_name} in {module_file}.\n"
+                                "By default, this should be a string literal. "
+                                "If dynamic evaluation is necessary, use "
+                                "`safe_import_context`."
+                            )
+                            def raise_err(self, msg=msg): raise ValueError(msg)
+                            ctx[target.id] = classproperty(raise_err)
         if isinstance(node, ast.FunctionDef):
             if node.name in known_methods:
                 ctx[node.name] = lambda *args, **kwargs: None
+
+    if ctx['name'] is None:
+        warnings.warn(
+            f"The class {cls_name} in module {module_file} has no name "
+            "attribute. Using the module filename as a fallback.",
+            UserWarning
+        )
+        ctx['name'] = name